progress on jsdom and dompurify

+++ src/routes.js
@@ -1,5 +1,13 @@
 import { ObjectId } from 'mongodb';
+import { JSDOM } from 'jsdom';
+import createDOMPurify from 'dompurify';
 
+// create headless browser
+const window = new JSDOM('').window;
+const DOMPurify = createDOMPurify(window);
+DOMPurify.setConfig({ ALLOWED_TAGS: [] });
+
+// confirm password
 const hasPassword = (req, res, next) => {
   if (req.params.password === process.env.PASSWORD) next();
   else res.status(401).json({ message: 'Unauthorized' });
@@ -8,10 +16,13 @@ const hasPassword = (req, res, next) => {
 export default (app, db) => {
   app.get('/api/resources/:password/:collection', hasPassword, async (req, res) => {
     try {
-      const { collection } = req.params;
+      let { collection } = req.params;
 
       // ensure clean param
       if (/^\$/.test(collection)) throw new Error('No topic with that name.');
+      collection = DOMPurify.sanitize(collection);
+
+      if (!collection) return res.status(400).json({ message: 'Bad Request' });
 
       // find all from collection and send
       const docs = await db.collection(collection).find().sort('createdAt', -1).toArray();
@@ -44,12 +55,28 @@ export default (app, db) => {
 
   app.post('/api/resource/:password/:collection', hasPassword, async (req, res) => {
     try {
-      const { collection } = req.params;
+      let { collection } = req.params;
 
       // ensure clean param
       if (/^\$/.test(collection)) throw new Error('No topic with that name.');
 
-      let { description, keywords, link, createdAt } = req.body;
+      // confirm type
+      let { createdAt } = req.body;
+      if (typeof createdAt !== 'number') return res.status(400).json({ message: 'Bad Request' });
+
+      // sanitize
+      collection = DOMPurify.sanitize(collection);
+      let description = DOMPurify.sanitize(req.body.description);
+      let link = DOMPurify.sanitize(req.body.link);
+      let keywords = req.body.keywords.reduce((arr, keyword) => {
+        keyword = DOMPurify.sanitize(keyword);
+        if (keyword) arr.push(keyword);
+        return arr;
+      }, []);
+
+      if (!collection || !description || !keywords.length || !link) {
+        return res.status(400).json({ message: 'Bad Request' });
+      }
 
       // insert doc
       const newDoc = await db
@@ -66,8 +93,16 @@ export default (app, db) => {
 
   app.post('/api/collection/:password', hasPassword, async (req, res) => {
     try {
+      let { collection } = req.body;
+
+      // ensure clean payload
+      if (/^\$/.test(collection)) throw new Error('No topic with that name.');
+      collection = DOMPurify.sanitize(collection);
+
+      if (!collection) return res.status(400).json({ message: 'Bad Request' });
+
       // create collection
-      const newCollection = await db.createCollection(req.body.collection);
+      const newCollection = await db.createCollection(collection);
       const newNamespace = newCollection.namespace.split('.')[1];
 
       // get all collection names
@@ -84,18 +119,29 @@ export default (app, db) => {
 
   app.put('/api/resource/:password/:collection/:id', hasPassword, async (req, res) => {
     try {
-      const { collection, id } = req.params;
+      let { collection, id } = req.params;
 
       // ensure clean params
       if (/^\$/.test(collection)) throw new Error('No topic with that name.');
       if (/^\$/.test(id)) throw new Error('No resource with that id.');
+      collection = DOMPurify.sanitize(collection);
+      id = DOMPurify.sanitize(id);
 
-      // ensure body only contains description, keywords, and link
+      if (!collection || !id) return res.status(400).json({ message: 'Bad Request' });
+
+      // ensure body only contains sanitized description, keywords, and link
       const body = {
-        description: req.body.description,
-        keywords: req.body.keywords,
-        link: req.body.link
+        description: DOMPurify.sanitize(req.body.description),
+        keywords: req.body.keywords.reduce((arr, keyword) => {
+          keyword = DOMPurify.sanitize(keyword);
+          if (keyword) arr.push(keyword);
+          return arr;
+        }, []),
+        link: DOMPurify.sanitize(req.body.link)
       };
+
+      if (!body.description || !body.keywords.length || !body.link)
+        return res.status(400).json({ message: 'Bad Request' });
 
       // find and update doc
       const updatedDoc = await db
@@ -121,11 +167,15 @@ export default (app, db) => {
 
   app.put('/api/resource/add-pin/:password/:collection/:id', hasPassword, async (req, res) => {
     try {
-      const { collection, id } = req.params;
+      let { collection, id } = req.params;
 
       // ensure clean params
       if (/^\$/.test(collection)) throw new Error('No topic with that name.');
       if (/^\$/.test(id)) throw new Error('No resource with that id.');
+      collection = DOMPurify.sanitize(collection);
+      id = DOMPurify.sanitize(id);
+
+      if (!collection || !id) return res.status(400).json({ message: 'Bad Request' });
 
       // query id and add isPinned: true
       const pinnedDoc = await db
@@ -150,11 +200,15 @@ export default (app, db) => {
 
   app.put('/api/resource/remove-pin/:password/:collection/:id', hasPassword, async (req, res) => {
     try {
-      const { collection, id } = req.params;
+      let { collection, id } = req.params;
 
       // ensure clean params
       if (/^\$/.test(collection)) throw new Error('No topic with that name.');
       if (/^\$/.test(id)) throw new Error('No resource with that id.');
+      collection = DOMPurify.sanitize(collection);
+      id = DOMPurify.sanitize(id);
+
+      if (!collection || !id) return res.status(400).json({ message: 'Bad Request' });
 
       // query id and remove isPinned: true
       const unpinnedDoc = await db
@@ -180,11 +234,17 @@ export default (app, db) => {
     hasPassword,
     async (req, res) => {
       try {
-        const { fromCollection, toCollection } = req.params;
+        let { fromCollection, toCollection } = req.params;
 
         // ensure clean params
-        if (/^\$/.test(fromCollection)) throw new Error('No topic with that name.');
-        if (/^\$/.test(toCollection)) throw new Error('Not a valid topic name.');
+        if (/^\$|/.test(fromCollection)) throw new Error('No topic with that name.');
+        if (/^\$|/.test(toCollection)) throw new Error('Not a valid topic name.');
+        fromCollection = DOMPurify.sanitize(fromCollection.trim());
+        toCollection = DOMPurify.sanitize(toCollection.trim());
+
+        if (!fromCollection || !toCollection) {
+          return res.status(400).json({ message: 'Bad Request' });
+        }
 
         // rename collection
         const updatedCollection = await db.renameCollection(fromCollection, toCollection);
@@ -203,11 +263,15 @@ export default (app, db) => {
 
   app.delete('/api/resource/:password/:collection/:id', hasPassword, async (req, res) => {
     try {
-      const { collection, id } = req.params;
+      let { collection, id } = req.params;
 
       // ensure clean params
       if (/^\$/.test(collection)) throw new Error('No topic with that name.');
       if (/^\$/.test(id)) throw new Error('No resource with that id.');
+      collection = DOMPurify.sanitize(collection);
+      id = DOMPurify.sanitize(id);
+
+      if (!collection || !id) return res.status(400).json({ message: 'Bad Request' });
 
       const deletedDoc = await db
         .collection(collection)
@@ -231,10 +295,13 @@ export default (app, db) => {
 
   app.delete('/api/collection/:password/:collection', hasPassword, async (req, res) => {
     try {
-      const { collection } = req.params;
+      let { collection } = req.params;
 
       // ensure clean param
       if (/^\$/.test(collection)) throw new Error('No topic with that name.');
+      collection = DOMPurify.sanitize(collection);
+
+      if (!collection) return res.status(400).json({ message: 'Bad Request' });
 
       // drop collection
       const dropped = await db.dropCollection(collection);
